+++ final_task/display_driver/spi_module.h
@@ -3,6 +3,6 @@
 #include <linux/types.h>
 
 int init_spi_wrapper(int spi_bus_index);
-int spi_wrapper_write(uint8_t *tx_buffer, uint16_t len);
+int spi_wrapper_write(uint8_t *tx_buffer, size_t len);
 void deinit_spi_wrapper(void);
 #endif
